More item modification
Inventory objects now MORE OR LESS work properly but get displayed "on
top of" each other, so something in the display code probably needs
modifying.

IMO the ideal approach would be to place the items @ regular intervals -
e.g. with one in inventory it's displayed at about (SCREEN_LENGTH/2);
with two, one is displayed 1/3 in and the other 2/3 in, etc.